Remove long delays. Fix some sending values. Update secrets. Turn on OTA. Move sensors classes to separate files.
