--- .vscode/launch.json
@@ -1,32 +1,0 @@
-// AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
-
-// PIO Unified Debugger
-//
-// Documentation: https://docs.platformio.org/page/plus/debugging.html
-// Configuration: https://docs.platformio.org/page/projectconf/section_env_debug.html
-
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "platformio-debug",
-            "request": "launch",
-            "name": "PIO Debug",
-            "executable": "c:/Dropbox/Development/Arduino/KResistance/Projects/K.Iot.ClimateStation/.pioenvs/nodemcuv2/firmware.elf",
-            "toolchainBinDir": "",
-            "preLaunchTask": {
-                "type": "PlatformIO",
-                "task": "Pre-Debug"
-            },
-            "internalConsoleOptions": "openOnSessionStart"
-        },
-        {
-            "type": "platformio-debug",
-            "request": "launch",
-            "name": "PIO Debug (skip Pre-Debug)",
-            "executable": "c:/Dropbox/Development/Arduino/KResistance/Projects/K.Iot.ClimateStation/.pioenvs/nodemcuv2/firmware.elf",
-            "toolchainBinDir": "",
-            "internalConsoleOptions": "openOnSessionStart"
-        }
-    ]
-}
