+++ platformio.ini
@@ -1,23 +1,10 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
 [env:nodemcuv2]
-platform = https://github.com/platformio/platform-espressif8266.git#feature/stage
+platform =  espressif8266
 board = nodemcuv2
 framework = arduino
 build_flags =
   !python -m secrets
 board_build.f_cpu = 160000000L
-upload_speed = 921600
-upload_port = COM5
-#upload_port = 192.168.1.210
 monitor_speed = 115200
 lib_deps =
   Adafruit Unified Sensor
@@ -26,3 +13,14 @@ lib_deps =
   DHT sensor library for ESPx
   K.MQSensors=https://github.com/KResistance/K.MQSensors.git#extensions
 lib_ldf_mode = deep+
+
+; COM port upload
+upload_port = COM4
+upload_speed = 921600
+board_build.flash_mode = dio
+
+; OTA upload
+; upload_port = name.local
+; upload_flags =
+;   --port=port
+;   --auth=pass
