+++ secrets.py.example
@@ -1,7 +1,12 @@
+location = "a"
+
 print(
-  "'-DWIFI_SSID=\"ssid\"'
-  '-DWIFI_PASS=\"pass\"'
-  '-DGRAPHITE_HOST=\"hostname\"'
-  '-DGRAPHITE_PORT=65000' \
-  '-DGRAPHITE_PREFIX=\"some.series.\"'"
+  "'-DMODULE_NAME=\"esp8266-" + location + "\"' \
+  '-DWIFI_SSID=\"ssid\"' \
+  '-DWIFI_PASS=\"pass\"' \
+  '-DOTA_PORT=8266' \
+  '-DOTA_PASS=\"ota_pass\"' \
+  '-DGRAPHITE_HOST=\"hostname\"' \
+  '-DGRAPHITE_PORT=80' \
+  '-DGRAPHITE_PREFIX=\"some.series." + location + ".\"'"
 )
